+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added array initialization syntax with `for` (not yet finalized).
 
+* Added `atascii` text codec.
+
 * Fixed several bugs, most importantly invalid offsets for branching instructions.
 
 ## 0.2.2